Refactoring from static helpers to injected service
