+++ google-cloud-tools-plugin/resources/META-INF/plugin.xml
@@ -59,6 +59,9 @@ Code inspections for App Engine Java code.</description>
                         serviceImplementation="com.google.cloud.tools.intellij.IdeaCloudToolsPluginInfoService"/>
     <applicationService serviceInterface="com.google.cloud.tools.intellij.CloudToolsPluginConfigurationService"
                         serviceImplementation="com.google.cloud.tools.intellij.DefaultCloudToolsPluginConfigurationService"/>
+    <applicationService serviceInterface="com.google.cloud.tools.intellij.appengine.cloud.AppEngineProjectHelper"
+                        serviceImplementation="com.google.cloud.tools.intellij.appengine.cloud.AppEngineProjectHelperImpl"/>
+
     <localInspection language="JAVA" shortName="ApiName" bundle="messages.EndpointBundle"  hasStaticDescription="true"
                      key="api.name.name" groupKey="inspections.group.name" enabledByDefault="true" level="ERROR"
                      implementationClass="com.google.cloud.tools.intellij.appengine.validation.ApiNameInspection"/>
